Reduced printing to avoid syslog spam.
Removed unecessary printf in lib/master.c and printk in ioctl.c to avoid
syslog spam. If ECAT_GetSlaveInfo in system SW is called on a slave that
does not exist you get a printf to stderr and printk in syslog. This is
however done every 10ms on each IO card in some applications. The
already returned error code is enough for proper error handling.